+ support Windows7

+++ features/debloatOneDrive/apply.ps1
@@ -1,6 +1,8 @@
 #Requires -RunAsAdministrator
 # reference: https://github.com/W4RH4WK/Debloat-Windows-10
 
+if ($osbver.Major -lt 7) { exit }
+
 Stop-Process -Name OneDrive -ea 0
 
 if (Test-Path "$env:LOCALAPPDATA\Microsoft\OneDrive\OneDrive.exe") {
